docs_v2/module_info_generator/requirements.txt
GitPython>=3.1.0
PyYAML>=6.0
pathlib
